backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
httpx==0.26.0
Pillow==10.2.0
sqlalchemy==2.0.25
aiosmtplib==3.0.1
pydantic[email]==2.5.3

# ML dependencies for CLIP
torch>=2.0.0
transformers==4.40.0
numpy>=1.24.0

# HEIC/HEIF image format support
pillow-heif>=0.15.0
